requirements.txt
blist==1.3.6
cassandra-driver==2.1.4
elasticsearch==1.4.0
futures==2.2.0
six==1.9.0
urllib3==1.10.2
